Separate decoder from capture.
