+++ src/capture.rs
@@ -1,5 +1,4 @@
 use std::ops::Range;
-use std::mem::size_of;
 
 use crate::file_vec::FileVec;
 use crate::hybrid_index::HybridIndex;
@@ -7,13 +6,8 @@ use crate::usb::{
     PID,
     PacketFields,
     SetupFields,
-    RequestType,
     Direction,
-    Recipient,
-    StandardRequest,
-    DescriptorType,
     DeviceDescriptor,
-    ConfigDescriptor,
     Configuration,
     ControlTransfer,
 };
@@ -55,43 +49,32 @@ bitfield! {
     #[derive(Copy, Clone, Debug, Default, Pod, Zeroable)]
     #[repr(C)]
     pub struct Endpoint(u64);
-    u64, device_id, set_device_id: 51, 0;
-    u8, device_address, set_device_address: 58, 52;
-    u8, number, set_number: 63, 59;
+    pub u64, device_id, set_device_id: 51, 0;
+    pub u8, device_address, set_device_address: 58, 52;
+    pub u8, number, set_number: 63, 59;
 }
 
 bitfield! {
     #[derive(Copy, Clone, Debug, Default, Pod, Zeroable)]
     #[repr(C)]
     pub struct TransferIndexEntry(u64);
-    u64, transfer_id, set_transfer_id: 51, 0;
-    u16, endpoint_id, set_endpoint_id: 62, 52;
-    u8, _is_start, _set_is_start: 63, 63;
+    pub u64, transfer_id, set_transfer_id: 51, 0;
+    pub u16, endpoint_id, set_endpoint_id: 62, 52;
+    pub u8, _is_start, _set_is_start: 63, 63;
 }
 
 impl TransferIndexEntry {
-    fn is_start(&self) -> bool {
+    pub fn is_start(&self) -> bool {
         self._is_start() != 0
     }
-    fn set_is_start(&mut self, value: bool) {
+    pub fn set_is_start(&mut self, value: bool) {
         self._set_is_start(value as u8)
     }
 }
 
-#[derive(Default)]
-struct TransactionState {
-    first: PID,
-    last: PID,
-    start: u64,
-    count: u64,
-    endpoint_id: usize,
-    setup: Option<SetupFields>,
-    payload: Vec<u8>,
-}
-
 #[derive(Copy, Clone, IntoPrimitive, FromPrimitive, PartialEq)]
 #[repr(u8)]
-enum EndpointState {
+pub enum EndpointState {
     #[default]
     Idle = 0,
     Starting = 1,
@@ -101,7 +84,7 @@ enum EndpointState {
 
 #[derive(Copy, Clone, Debug, FromPrimitive)]
 #[repr(u8)]
-enum EndpointType {
+pub enum EndpointType {
     Control       = 0x00,
     Isochronous   = 0x01,
     Bulk          = 0x02,
@@ -112,28 +95,21 @@ enum EndpointType {
     Invalid       = 0x11,
 }
 
-struct EndpointData {
-    device_id: usize,
-    number: usize,
-    transaction_ids: HybridIndex,
-    transfer_index: HybridIndex,
-    transaction_start: u64,
-    transaction_count: u64,
-    last: PID,
-    setup: Option<SetupFields>,
-    payload: Vec<u8>,
+pub struct EndpointTraffic {
+    pub transaction_ids: HybridIndex,
+    pub transfer_index: HybridIndex,
 }
 
-struct DeviceData {
-    device_descriptor: Option<DeviceDescriptor>,
-    configurations: Vec<Option<Configuration>>,
-    configuration_id: Option<usize>,
-    endpoint_types: Vec<EndpointType>,
-    strings: Vec<Option<Vec<u8>>>,
+pub struct DeviceData {
+    pub device_descriptor: Option<DeviceDescriptor>,
+    pub configurations: Vec<Option<Configuration>>,
+    pub configuration_id: Option<usize>,
+    pub endpoint_types: Vec<EndpointType>,
+    pub strings: Vec<Option<Vec<u8>>>,
 }
 
 impl DeviceData {
-    fn endpoint_type(&self, number: usize) -> EndpointType {
+    pub fn endpoint_type(&self, number: usize) -> EndpointType {
         use EndpointType::*;
         match number {
             0 => Control,
@@ -143,7 +119,7 @@ impl DeviceData {
         }
     }
 
-    fn update_endpoint_types(&mut self) {
+    pub fn update_endpoint_types(&mut self) {
         match self.configuration_id {
             Some(id) => match &self.configurations[id] {
                 Some(config) => {
@@ -163,26 +139,18 @@ impl DeviceData {
     }
 }
 
-const USB_MAX_DEVICES: usize = 128;
-const USB_MAX_ENDPOINTS: usize = 16;
-
 pub struct Capture {
-    item_index: HybridIndex,
-    packet_index: HybridIndex,
-    packet_data: FileVec<u8>,
-    transaction_index: HybridIndex,
-    transfer_index: FileVec<TransferIndexEntry>,
-    device_index: [i8; USB_MAX_DEVICES],
-    devices: FileVec<Device>,
-    device_data: Vec<DeviceData>,
-    endpoint_index: [[i16; USB_MAX_ENDPOINTS]; USB_MAX_DEVICES],
-    endpoints: FileVec<Endpoint>,
-    endpoint_data: Vec<EndpointData>,
-    endpoint_states: FileVec<u8>,
-    endpoint_state_index: HybridIndex,
-    last_endpoint_state: Vec<u8>,
-    last_item_endpoint: i16,
-    transaction_state: TransactionState,
+    pub item_index: HybridIndex,
+    pub packet_index: HybridIndex,
+    pub packet_data: FileVec<u8>,
+    pub transaction_index: HybridIndex,
+    pub transfer_index: FileVec<TransferIndexEntry>,
+    pub devices: FileVec<Device>,
+    pub device_data: Vec<DeviceData>,
+    pub endpoints: FileVec<Endpoint>,
+    pub endpoint_traffic: Vec<EndpointTraffic>,
+    pub endpoint_states: FileVec<u8>,
+    pub endpoint_state_index: HybridIndex,
 }
 
 impl Default for Capture {
@@ -206,78 +174,6 @@ impl Transaction {
         match &self.payload_byte_range {
             Some(range) => Some(range.end - range.start),
             None => None
-        }
-    }
-}
-
-#[derive(PartialEq)]
-enum DecodeStatus {
-    NEW,
-    CONTINUE,
-    DONE,
-    INVALID
-}
-
-impl TransactionState {
-    pub fn status(&mut self, packet: &[u8]) -> DecodeStatus {
-        let next = PID::from(packet[0]);
-        use PID::*;
-        match (self.first, self.last, next) {
-
-            // SETUP, IN or OUT always start a new transaction.
-            (_, _, SETUP | IN | OUT) => DecodeStatus::NEW,
-
-            // SOF when there is no existing transaction starts a new
-            // "transaction" representing an idle period on the bus.
-            (_, Malformed, SOF) => DecodeStatus::NEW,
-            // Additional SOFs extend this "transaction", more may follow.
-            (_, SOF, SOF) => DecodeStatus::CONTINUE,
-
-            // SETUP must be followed by DATA0.
-            (_, SETUP, DATA0) => {
-                // The packet must have the correct size.
-                match packet.len() {
-                    11 => {
-                        self.setup = Some(
-                            SetupFields::from_data_packet(packet));
-                        // Wait for ACK.
-                        DecodeStatus::CONTINUE
-                    },
-                    _ => DecodeStatus::INVALID
-                }
-            }
-            // ACK then completes the transaction.
-            (SETUP, DATA0, ACK) => DecodeStatus::DONE,
-
-            // IN may be followed by NAK or STALL, completing transaction.
-            (_, IN, NAK | STALL) => DecodeStatus::DONE,
-            // IN or OUT may be followed by DATA0 or DATA1, wait for status.
-            (_, IN | OUT, DATA0 | DATA1) => {
-                if packet.len() >= 3 {
-                    let range = 1 .. (packet.len() - 2);
-                    self.payload = packet[range].to_vec();
-                    DecodeStatus::CONTINUE
-                } else {
-                    DecodeStatus::INVALID
-                }
-            },
-            // An ACK or NYET then completes the transaction.
-            (IN | OUT, DATA0 | DATA1, ACK | NYET) => DecodeStatus::DONE,
-            // OUT may also be completed by NAK or STALL.
-            (OUT, DATA0 | DATA1, NAK | STALL) => DecodeStatus::DONE,
-
-            // Any other case is not a valid part of a transaction.
-            _ => DecodeStatus::INVALID,
-        }
-    }
-
-    fn completed(&self) -> bool {
-        use PID::*;
-        // A transaction is completed if it has 3 valid packets and is
-        // acknowledged with an ACK or NYET handshake.
-        match (self.count, self.last) {
-            (3, ACK | NYET) => true,
-            (..)            => false
         }
     }
 }
@@ -321,33 +217,19 @@ pub fn fmt_index(idx: &HybridIndex) -> String {
 
 impl Capture {
     pub fn new() -> Self {
-        let mut capture = Capture {
+        Capture {
             item_index: HybridIndex::new(1).unwrap(),
             packet_index: HybridIndex::new(2).unwrap(),
             packet_data: FileVec::new().unwrap(),
             transaction_index: HybridIndex::new(1).unwrap(),
             transfer_index: FileVec::new().unwrap(),
-            device_index: [-1; USB_MAX_DEVICES],
             devices: FileVec::new().unwrap(),
             device_data: Vec::new(),
             endpoints: FileVec::new().unwrap(),
-            endpoint_data: Vec::new(),
-            endpoint_index: [[-1; USB_MAX_ENDPOINTS]; USB_MAX_DEVICES],
+            endpoint_traffic: Vec::new(),
             endpoint_states: FileVec::new().unwrap(),
             endpoint_state_index: HybridIndex::new(1).unwrap(),
-            last_endpoint_state: Vec::new(),
-            last_item_endpoint: -1,
-            transaction_state: TransactionState::default(),
-        };
-        capture.add_endpoint(0, EndpointType::Invalid as usize);
-        capture.add_endpoint(0, EndpointType::Framing as usize);
-        capture
-    }
-
-    pub fn handle_raw_packet(&mut self, packet: &[u8]) {
-        self.transaction_update(packet);
-        self.packet_index.push(self.packet_data.len()).unwrap();
-        self.packet_data.append(packet).unwrap();
+        }
     }
 
     pub fn print_storage_summary(&self) {
@@ -363,13 +245,13 @@ impl Capture {
         let mut xfr_count = 0;
         let mut xfr_entries = 0;
         let mut xfr_size = 0;
-        for ep_data in &self.endpoint_data {
-            trx_count += ep_data.transaction_ids.len();
-            trx_entries += ep_data.transaction_ids.entry_count();
-            trx_size += ep_data.transaction_ids.size();
-            xfr_count += ep_data.transfer_index.len();
-            xfr_entries += ep_data.transfer_index.entry_count();
-            xfr_size += ep_data.transfer_index.size();
+        for ep_traf in &self.endpoint_traffic {
+            trx_count += ep_traf.transaction_ids.len();
+            trx_entries += ep_traf.transaction_ids.entry_count();
+            trx_size += ep_traf.transaction_ids.size();
+            xfr_count += ep_traf.transfer_index.len();
+            xfr_entries += ep_traf.transfer_index.entry_count();
+            xfr_size += ep_traf.transfer_index.size();
             overhead += trx_size + xfr_size;
         }
         let ratio = (overhead as f32) / (self.packet_data.size() as f32);
@@ -397,369 +279,6 @@ impl Capture {
         )
     }
 
-    fn transaction_update(&mut self, packet: &[u8]) {
-        let pid = PID::from(packet[0]);
-        match self.transaction_state.status(packet) {
-            DecodeStatus::NEW => {
-                self.transaction_end();
-                self.transaction_start(packet);
-            },
-            DecodeStatus::CONTINUE => {
-                self.transaction_append(pid);
-            },
-            DecodeStatus::DONE => {
-                self.transaction_append(pid);
-                self.transaction_end();
-            },
-            DecodeStatus::INVALID => {
-                self.transaction_end();
-                self.transaction_start(packet);
-                self.transaction_end();
-            },
-        };
-    }
-
-    fn transaction_start(&mut self, packet: &[u8]) {
-        let state = &mut self.transaction_state;
-        state.start = self.packet_index.len();
-        state.count = 1;
-        state.first = PID::from(packet[0]);
-        state.last = state.first;
-        match PacketFields::from_packet(&packet) {
-            PacketFields::SOF(_) => {
-                self.transaction_state.endpoint_id = 1;
-            },
-            PacketFields::Token(token) => {
-                let addr = token.device_address() as usize;
-                let num = token.endpoint_number() as usize;
-                if self.endpoint_index[addr][num] < 0 {
-                    let endpoint_id = self.endpoints.len() as i16;
-                    self.endpoint_index[addr][num] = endpoint_id;
-                    self.add_endpoint(addr, num);
-                }
-                self.transaction_state.endpoint_id =
-                    self.endpoint_index[addr][num] as usize;
-            },
-            _ => {
-                self.transaction_state.endpoint_id = 0;
-            }
-        }
-    }
-
-    fn transaction_append(&mut self, pid: PID) {
-        let state = &mut self.transaction_state;
-        state.count += 1;
-        state.last = pid;
-    }
-
-    fn transaction_end(&mut self) {
-        self.add_transaction();
-        let state = &mut self.transaction_state;
-        state.count = 0;
-        state.first = PID::Malformed;
-        state.last = PID::Malformed;
-        state.setup = None;
-    }
-
-    fn add_transaction(&mut self) {
-        if self.transaction_state.count == 0 { return }
-        self.transfer_update();
-        self.transaction_index.push(self.transaction_state.start).unwrap();
-    }
-
-    fn add_endpoint(&mut self, addr: usize, num: usize) {
-        if self.device_index[addr] == -1 {
-            self.device_index[addr] = self.devices.size() as i8;
-            let device = Device { address: addr as u8 };
-            self.devices.push(&device).unwrap();
-            let dev_data = DeviceData {
-                device_descriptor: None,
-                configurations: Vec::new(),
-                configuration_id: None,
-                endpoint_types: vec![
-                    EndpointType::Unidentified; USB_MAX_ENDPOINTS],
-                strings: Vec::new(),
-            };
-            self.device_data.push(dev_data);
-        }
-        let ep_data = EndpointData {
-            number: num as usize,
-            device_id: self.device_index[addr] as usize,
-            transaction_ids: HybridIndex::new(1).unwrap(),
-            transfer_index: HybridIndex::new(1).unwrap(),
-            transaction_start: 0,
-            transaction_count: 0,
-            last: PID::Malformed,
-            setup: None,
-            payload: Vec::new(),
-        };
-        self.endpoint_data.push(ep_data);
-        let mut endpoint = Endpoint::default();
-        endpoint.set_device_id(self.device_index[addr] as u64);
-        endpoint.set_device_address(addr as u8);
-        endpoint.set_number(num as u8);
-        self.endpoints.push(&endpoint).unwrap();
-        self.last_endpoint_state.push(EndpointState::Idle as u8);
-    }
-
-    fn decode_request(&mut self) {
-        let endpoint_id = self.transaction_state.endpoint_id;
-        let ep_data = &self.endpoint_data[endpoint_id];
-        let fields = ep_data.setup.as_ref().unwrap();
-        let req_type = fields.type_fields.request_type();
-        let request = StandardRequest::from(fields.request);
-        match (req_type, request) {
-            (RequestType::Standard, StandardRequest::GetDescriptor)
-                => self.decode_descriptor_read(),
-            (RequestType::Standard, StandardRequest::SetConfiguration)
-                => self.decode_configuration_set(),
-            (..) => {}
-        }
-    }
-
-    fn decode_descriptor_read(&mut self) {
-        let endpoint_id = self.transaction_state.endpoint_id;
-        let ep_data = &mut self.endpoint_data[endpoint_id];
-        let fields = ep_data.setup.as_ref().unwrap();
-        let recipient = fields.type_fields.recipient();
-        let desc_type = DescriptorType::from((fields.value >> 8) as u8);
-        let payload = &ep_data.payload;
-        let length = payload.len();
-        match (recipient, desc_type) {
-            (Recipient::Device, DescriptorType::Device) => {
-                if length == size_of::<DeviceDescriptor>() {
-                    let device_id = ep_data.device_id;
-                    let dev_data = &mut self.device_data[device_id];
-                    dev_data.device_descriptor =
-                        Some(DeviceDescriptor::from_bytes(payload));
-                }
-            },
-            (Recipient::Device, DescriptorType::Configuration) => {
-                let size = size_of::<ConfigDescriptor>();
-                if length >= size {
-                    let device_id = ep_data.device_id;
-                    let dev_data = &mut self.device_data[device_id];
-                    let configurations = &mut dev_data.configurations;
-                    let configuration = Configuration::from_bytes(&payload);
-                    if let Some(config) = configuration {
-                        let config_id =
-                            config.descriptor.config_value as usize;
-                        while configurations.len() <= config_id {
-                            configurations.push(None);
-                        }
-                        configurations[config_id] = Some(config);
-                        dev_data.update_endpoint_types();
-                    }
-                }
-            },
-            (Recipient::Device, DescriptorType::String) => {
-                if length >= 2 {
-                    let device_id = ep_data.device_id;
-                    let strings = &mut self.device_data[device_id].strings;
-                    let string_id = (fields.value & 0xFF) as usize;
-                    while strings.len() <= string_id {
-                        strings.push(None);
-                    }
-                    strings[string_id] = Some(payload[2..length].to_vec());
-                }
-            },
-            _ => {}
-        }
-    }
-
-    fn decode_configuration_set(&mut self) {
-        let endpoint_id = self.transaction_state.endpoint_id;
-        let ep_data = &mut self.endpoint_data[endpoint_id];
-        let device_id = ep_data.device_id;
-        let dev_data = &mut self.device_data[device_id];
-        let fields = ep_data.setup.as_ref().unwrap();
-        let config_id = fields.value as usize;
-        dev_data.configuration_id = Some(config_id);
-        dev_data.update_endpoint_types();
-    }
-
-    fn transfer_status(&mut self) -> DecodeStatus {
-        let next = self.transaction_state.first;
-        let endpoint_id = self.transaction_state.endpoint_id;
-        let ep_data = &mut self.endpoint_data[endpoint_id];
-        let dev_data = &self.device_data[ep_data.device_id];
-        let ep_type = &dev_data.endpoint_type(ep_data.number);
-        use PID::*;
-        use EndpointType::*;
-        use Direction::*;
-        match (ep_type, ep_data.last, next) {
-
-            // A SETUP transaction starts a new control transfer.
-            // Store the setup fields to interpret the request.
-            (Control, _, SETUP) => {
-                ep_data.setup = self.transaction_state.setup.take();
-                DecodeStatus::NEW
-            },
-
-            (Control, _, _) => match &ep_data.setup {
-                // No control transaction is valid unless setup was done.
-                None => DecodeStatus::INVALID,
-                // If setup was done then valid transactions depend on the
-                // contents of the setup data packet.
-                Some(fields) => {
-                    let with_data = fields.length != 0;
-                    let direction = fields.type_fields.direction();
-                    match (direction, with_data, ep_data.last, next) {
-
-                        // If there is data to transfer, setup stage is
-                        // followed by IN/OUT at data stage in the direction
-                        // of the request. IN/OUT may then be repeated.
-                        (In,  true, SETUP, IN ) |
-                        (Out, true, SETUP, OUT) |
-                        (In,  true, IN,    IN ) |
-                        (Out, true, OUT,   OUT) => {
-                            if self.transaction_state.completed() {
-                                ep_data.payload.extend(
-                                    &self.transaction_state.payload);
-                            }
-                            // Await status stage.
-                            DecodeStatus::CONTINUE
-                        },
-
-                        // If there is no data to transfer, setup stage is
-                        // followed by IN/OUT at status stage in the opposite
-                        // direction to the request. If there is data, then
-                        // the status stage follows the data stage.
-                        (In,  false, SETUP, OUT) |
-                        (Out, false, SETUP, IN ) |
-                        (In,  true,  IN,    OUT) |
-                        (Out, true,  OUT,   IN ) => {
-                            self.decode_request();
-                            DecodeStatus::DONE
-                        },
-                        // Any other sequence is invalid.
-                        (..) => DecodeStatus::INVALID
-                    }
-                }
-            },
-
-            // An IN or OUT transaction on a non-control endpoint,
-            // with no transfer in progress, starts a new transfer.
-            (_, Malformed, IN | OUT) => DecodeStatus::NEW,
-
-            // IN or OUT may then be repeated.
-            (_, IN, IN) => DecodeStatus::CONTINUE,
-            (_, OUT, OUT) => DecodeStatus::CONTINUE,
-
-            // A SOF group starts a special transfer, unless
-            // one is already in progress.
-            (Framing, Malformed, SOF) => DecodeStatus::NEW,
-
-            // Further SOF groups continue this transfer.
-            (Framing, SOF, SOF) => DecodeStatus::CONTINUE,
-
-            // Any other case is not a valid part of a transfer.
-            _ => DecodeStatus::INVALID
-        }
-    }
-
-    fn transfer_update(&mut self) {
-        let status = self.transfer_status();
-        let endpoint_id = self.transaction_state.endpoint_id;
-        let ep_data = &mut self.endpoint_data[endpoint_id];
-        let retry_needed =
-            ep_data.transaction_count > 0 &&
-            status != DecodeStatus::INVALID &&
-            !self.transaction_state.completed();
-        if retry_needed {
-            self.transfer_append(false);
-            return
-        }
-        match status {
-            DecodeStatus::NEW => {
-                self.transfer_end();
-                self.transfer_start();
-                self.transfer_append(true);
-            },
-            DecodeStatus::CONTINUE => {
-                self.transfer_append(true);
-            },
-            DecodeStatus::DONE => {
-                self.transfer_append(true);
-                self.transfer_end();
-            },
-            DecodeStatus::INVALID => {
-                self.transfer_end();
-                self.transfer_start();
-                self.transfer_append(false);
-                self.transfer_end();
-            }
-        }
-    }
-
-    fn transfer_start(&mut self) {
-        self.item_index.push(self.transfer_index.len()).unwrap();
-        let endpoint_id = self.transaction_state.endpoint_id;
-        self.last_item_endpoint = endpoint_id as i16;
-        self.add_transfer_entry(endpoint_id, true);
-        let ep_data = &mut self.endpoint_data[endpoint_id];
-        ep_data.transaction_start = ep_data.transaction_ids.len();
-        ep_data.transaction_count = 0;
-        ep_data.transfer_index.push(ep_data.transaction_start).unwrap();
-    }
-
-    fn transfer_append(&mut self, success: bool) {
-        let endpoint_id = self.transaction_state.endpoint_id;
-        let ep_data = &mut self.endpoint_data[endpoint_id];
-        ep_data.transaction_ids.push(self.transaction_index.len()).unwrap();
-        ep_data.transaction_count += 1;
-        if success {
-            ep_data.last = self.transaction_state.first;
-        }
-    }
-
-    fn transfer_end(&mut self) {
-        let endpoint_id = self.transaction_state.endpoint_id;
-        let ep_data = &self.endpoint_data[endpoint_id];
-        if ep_data.transaction_count > 0 {
-            if self.last_item_endpoint != (endpoint_id as i16) {
-                self.item_index.push(self.transfer_index.len()).unwrap();
-                self.last_item_endpoint = endpoint_id as i16;
-            }
-            self.add_transfer_entry(endpoint_id, false);
-        }
-        let ep_data = &mut self.endpoint_data[endpoint_id];
-        ep_data.transaction_count = 0;
-        ep_data.last = PID::Malformed;
-        ep_data.payload.clear();
-    }
-
-    fn add_transfer_entry(&mut self, endpoint_id: usize, start: bool) {
-        let ep_data = &mut self.endpoint_data[endpoint_id];
-        let mut entry = TransferIndexEntry::default();
-        entry.set_endpoint_id(endpoint_id as u16);
-        entry.set_transfer_id(ep_data.transfer_index.len());
-        entry.set_is_start(start);
-        self.transfer_index.push(&entry).unwrap();
-        self.add_endpoint_state(endpoint_id, start);
-    }
-
-    fn add_endpoint_state(&mut self, endpoint_id: usize, start: bool) {
-        let endpoint_count = self.endpoints.len() as usize;
-        for i in 0..endpoint_count {
-            use EndpointState::*;
-            self.last_endpoint_state[i] = {
-                let same = endpoint_id == i;
-                let last = EndpointState::from(self.last_endpoint_state[i]);
-                match (same, start, last) {
-                    (true, true,  _)               => Starting,
-                    (true, false, _)               => Ending,
-                    (false, _, Starting | Ongoing) => Ongoing,
-                    (false, _, Ending | Idle)      => Idle,
-                }
-            } as u8;
-        }
-        let last_state = self.last_endpoint_state.as_slice();
-        let state_offset = self.endpoint_states.len();
-        self.endpoint_states.append(last_state).unwrap();
-        self.endpoint_state_index.push(state_offset).unwrap();
-    }
-
     pub fn get_item(&mut self, parent: &Option<Item>, index: u64) -> Item {
         match parent {
             None => Item::Transfer(self.item_index.get(index).unwrap()),
@@ -775,9 +294,9 @@ impl Capture {
                     let entry = self.transfer_index.get(*transfer_index_id).unwrap();
                     let endpoint_id = entry.endpoint_id() as usize;
                     let transfer_id = entry.transfer_id();
-                    let ep_data = &mut self.endpoint_data[endpoint_id];
-                    let offset = ep_data.transfer_index.get(transfer_id).unwrap();
-                    ep_data.transaction_ids.get(offset + index).unwrap()
+                    let ep_traf = &mut self.endpoint_traffic[endpoint_id];
+                    let offset = ep_traf.transfer_index.get(transfer_id).unwrap();
+                    ep_traf.transaction_ids.get(offset + index).unwrap()
                 }),
             Transaction(transfer_index_id, transaction_id) =>
                 Packet(*transfer_index_id, *transaction_id, {
@@ -793,9 +312,9 @@ impl Capture {
                 let entry = self.transfer_index.get(*transfer_index_id).unwrap();
                 let endpoint_id = entry.endpoint_id() as usize;
                 let transfer_id = entry.transfer_id();
-                let ep_data = &mut self.endpoint_data[endpoint_id];
-                get_index_range(&mut ep_data.transfer_index,
-                    ep_data.transaction_ids.len(), transfer_id)
+                let ep_traf = &mut self.endpoint_traffic[endpoint_id];
+                get_index_range(&mut ep_traf.transfer_index,
+                    ep_traf.transaction_ids.len(), transfer_id)
             },
             Transaction(_, transaction_id) => {
                 get_index_range(&mut self.transaction_index,
@@ -929,13 +448,13 @@ impl Capture {
         let endpoint_state = self.get_endpoint_state(*transfer_index_id);
         let state_length = endpoint_state.len();
         let extended = self.transfer_extended(endpoint_id, *transfer_index_id);
-        let ep_data = &mut self.endpoint_data[endpoint_id];
+        let ep_traf = &mut self.endpoint_traffic[endpoint_id];
         let last_transaction = match item {
             Transaction(_, transaction_id) | Packet(_, transaction_id, _) => {
-                let range = get_index_range(&mut ep_data.transfer_index,
-                    ep_data.transaction_ids.len(), entry.transfer_id());
+                let range = get_index_range(&mut ep_traf.transfer_index,
+                    ep_traf.transaction_ids.len(), entry.transfer_id());
                 let last_transaction_id =
-                    ep_data.transaction_ids.get(range.end - 1).unwrap();
+                    ep_traf.transaction_ids.get(range.end - 1).unwrap();
                 *transaction_id == last_transaction_id
             }, _ => false
         };
@@ -1075,9 +594,9 @@ impl Capture {
                             endpoint_id: u16,
                             range: Range<u64>) -> ControlTransfer
     {
-        let ep_data = &mut self.endpoint_data[endpoint_id as usize];
+        let ep_traf = &mut self.endpoint_traffic[endpoint_id as usize];
         let transaction_ids =
-            ep_data.transaction_ids.get_range(range).unwrap();
+            ep_traf.transaction_ids.get_range(range).unwrap();
         let setup_transaction_id = transaction_ids[0];
         let setup_packet_id =
             self.transaction_index.get(setup_transaction_id)
@@ -1270,6 +789,7 @@ mod tests {
     use super::*;
     use std::fs::File;
     use std::io::{BufReader, BufWriter, BufRead, Write};
+    use crate::decoder::Decoder;
 
     fn write_item(cap: &mut Capture, item: &Item, depth: u8,
                   writer: &mut dyn Write)
@@ -1303,8 +823,9 @@ mod tests {
                 {
                     let mut pcap = pcap::Capture::from_file(cap_path).unwrap();
                     let mut cap = Capture::new();
+                    let mut decoder = Decoder::new(&mut cap);
                     while let Ok(packet) = pcap.next() {
-                        cap.handle_raw_packet(&packet);
+                        decoder.handle_raw_packet(&packet);
                     }
                     let out_file = File::create(out_path.clone()).unwrap();
                     let mut out_writer = BufWriter::new(out_file);
