+++ src/main.rs
@@ -27,6 +27,9 @@ use expander::ExpanderWrapper;
 mod capture;
 use capture::Capture;
 
+mod decoder;
+use decoder::Decoder;
+
 mod file_vec;
 mod hybrid_index;
 mod usb;
@@ -145,8 +148,9 @@ fn main() {
     let args: Vec<_> = std::env::args().collect();
     let mut pcap = pcap::Capture::from_file(&args[1]).unwrap();
     let mut cap = Capture::new();
+    let mut decoder = Decoder::new(&mut cap);
     while let Ok(packet) = pcap.next() {
-        cap.handle_raw_packet(&packet);
+        decoder.handle_raw_packet(&packet);
     }
     cap.print_storage_summary();
     let capture = Arc::new(Mutex::new(cap));
